Handle case for commits not touching BCI packages
The bot can create a commit (e.g. changing .obs/workflows.yml) that will not
touch any BCI packages. Such a commit will have a diff, but it will not result
in any package changes and thus it makes no sense to create the staging project.

Also, the reduce assertion would fail in that case, as an initial value was
missing.