reduce busybox image size by removing Index.db
Index.db is being created on demand by rpm(1) (as long as it is
run as root) so we don't strictly have to ship it. currently the
Index.db has a lot of empty space within which we could optimize away.

until then removing the file saves 10% of the total compressed size.