Fix commit range getter: exclude ancestor_ref
We use this function to get all commits between two refs, but we always want to
check everything between an ancestor and HEAD excluding the ancestor (the
ancestor is most frequently the last merge to the deployment branch). Therefore
we switch from returning a list and use a set instead where we remove the
ancestor commit.