Add a prefix into the release portion of the tag for stability-tags
The stability-tags feature in containers. may result in the generation of
identical release numbers for the same version from two different package
containers, such as "oldstable" and "stable."

When a new version is committed, the check-in counter and the rebuild
counters are reset to "1", resulting in a release number like 1.1.
Subsequent changes will have release numbers like 1.2 (rebuild on the same source)
and 2.1 (new source change without version change).

Here's an example:
  lang-stable: 1.70-1.1 (first build of the very first commit after version update)
  lang-oldstable: 1.69-5.1 (first build of the fifth source change after version update)

After a version rollover occurs from "stable" to "oldstable", the release numbers become:

  lang-oldstable: 1.70-1.1
  lang-stable: 1.71-1.1

Now there is a conflict with the tags that were previously produced by the lang-stable
container (see lang-stable-1.70-1.1 above).

In order To resolve this, the solution is to namespace
the tags with a prefix based on the stability ordering. This ensures that we have:

  lang-stable: 1.70-1.1.1
  lang-oldstable: 1.69-2.5.1

After a version rollover, the numbers now become:

  lang-oldstable: 1.70-2.1.1
  lang-stable: 1.71-1.1.1

To avoid conflicts, the tags are deconflicted based on the stability ordering.

This ensures that the same version in "oldstable" stability can not
generate the same release number than in "stable" stability.