Redefine the meaning of tag_version to exclude build_flavor
for build_flavor containers, we had previously to include the
build_flavor into the tag_version to disambiguate the tags and
buildnames. Change the meaning of tag_version to be without build_flavor
and instead explicitly handle build_flavor where necessary.

This commit mixes also a hook in Registry to use version, rather than
tag_version for the build_version, which is a must-have for the
AppCollection integration.

Add tests for ApplicationStackContainers in unit test, which was
entirely untested so far.