Introduce a branch_version concept in BCI
Now that we have build_flavors, we can add branch_versions, describing
the full hierachy:

Each container can have a single `name`, one or many `branch_versions`
and each of them one or many `build_flavors`.

this means our naming should become

{name}-{branch_version}-{build_flavor}

Introduce this for the AppCollection containers first.  This should
eventually replace tag_version, which is a broken concept as it
mixes namings, versions and flavors into and hence makes everything
convoluted.

This commit mixes also a hook in Registry to use version, rather than
tag_version for the build_version, which is a must-have for the
AppCollection integration.

Add tests for ApplicationStackContainers in unit test, which was
entirely untested so far.

Removal of the deprecated tag_version will happen in a later followup.