Limit the go version to three components
The go-openssl variant has/had a four component version which confuses the heck
out of the go ecosystem. Therefore jfkw patched go-openssl to no longer output
the fourth component, but now we have a mismatch between the output go-openssl
version and the go-openssl rpm package version.

By setting the replace_using_package_version service to parse the patch version,
we hopefully eliminate this discrepancy in the `GOLANG_VERSION` environment
variable.