[bot] Reimplement _get_commit_range_between_refs using git rev-list
The `_get_commit_range_between_refs` function was relying on a buggy recursion
which broke for unknown reasons and had ugly edge cases. Fortunately,
git-rev-list features more or less exactly the functionality that we need (with
a few added quircks, like occasionally including commits that are not children
of the ancestor). Therfore we move the function _get_commit_range_between_refs
into a new module, use git-rev-list instead of our recursion and add tests for
it.