Merge versioning behavior between Dockerfiles and Kiwi type builds
We had two versioning implementations, one for dockerfiles and one for
kiwi builds. So it happened what happens whenever there are two
implementations: they diverge in an incompatible way. So merge them
by deriving the kiwi_version from the build_version.

- 15.<SP>.0 for OsContainers (whatever type) on SLE15
- 16.0.0 for OsContainers (whatever type) on SLE16

We still skip setting a build-version for tumbleweed for anything
other than OsContainer/KIWI types per default to remain compatible
with the previous behavior. Although this is probably a bad idea.