Allow skipping adding the buildflavor in buildtags
build_flavor can be used to designate branches of the same package
(in which case adding a flavor-suffix to the tag makes sense) or can
be used, like in the kubevirt case, to share multiple containers that
are individually already having a flavor in the repository tag.

Add a toggle for that.