Remove last password change from /etc/shadow
The last password change field in /etc/shadow changes with every rebuild
because the file is generated using the current installation time.

Because this file is created after the installation, there's no easy way
to use SOURCE_DATE_EPOCH without affecting the package itself.

Since we don't set any passwords for any users, it is safe to remove it
in most cases as the accounts are system accounts, either locked or with
no usable password.